Remove broken spec
The behavior has changed since feac7c8a. When setting the result, if the
current state of the object is valid, it will remain valid indefinitely.
So testing this behavior now requires checking for validity first, then
setting the result. But this isn't any different than just having an
error on the object, so there's no reason for this test.